requirements.txt
Django==5.2.8
mysqlclient>=2.2.0
Pillow>=10.0.0
openpyxl>=3.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
